+++ tests/test_class_basics.rs
@@ -244,3 +244,47 @@ fn panic_unsendable_base() {
 fn panic_unsendable_child() {
     test_unsendable::<UnsendableChild>().unwrap();
 }
+
+fn get_length(obj: &PyAny) -> PyResult<usize> {
+    let length = obj.len()?;
+
+    Ok(length)
+}
+
+#[pyclass]
+struct ClassWithFromPyWithMethods {}
+
+#[pymethods]
+impl ClassWithFromPyWithMethods {
+    fn instance_method(&self, #[pyo3(from_py_with = "get_length")] argument: usize) -> usize {
+        argument
+    }
+    #[classmethod]
+    fn classmethod(_cls: &PyType, #[pyo3(from_py_with = "PyAny::len")] argument: usize) -> usize {
+        argument
+    }
+
+    #[staticmethod]
+    fn staticmethod(#[pyo3(from_py_with = "get_length")] argument: usize) -> usize {
+        argument
+    }
+}
+
+#[test]
+fn test_pymethods_from_py_with() {
+    Python::with_gil(|py| {
+        let instance = Py::new(py, ClassWithFromPyWithMethods {}).unwrap();
+
+        py_run!(
+            py,
+            instance,
+            r#"
+        arg = {1: 1, 2: 3}
+
+        assert instance.instance_method(arg) == 2
+        assert instance.classmethod(arg) == 2
+        assert instance.staticmethod(arg) == 2
+        "#
+        );
+    })
+}